rename additions.cpp->main.cpp; deleted old main.cpp; uncomment main() func in new main.cpp; now it's only 1 .cpp file in project
